requirements.txt
pytest
rich
asyncio
argparse
rich_argparse
patchright
Quart
Faker
pyngrok
requests
python-dotenv
